更新版本1.0/agent-fastapi/pyproject.toml
[project]
name = "agent-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dashscope>=1.24.9",
    "fastapi>=0.120.0",
    "httpx>=0.28.1",
    "langchain>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-mcp-adapters>=0.1.12",
    "langchain-openai>=1.1.0",
    "langgraph>=1.0.2",
    "langgraph-checkpoint-postgres>=3.0.0",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary,pool]>=3.2.12",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "pywin32>=311",
    "qianfan>=0.4.12.3",
    "sqlmodel>=0.0.27",
    "uvicorn[standard]>=0.38.0",
    "websockets>=15.0.1",
]
